+++ src/main/resources/templates/common/_footer.html
@@ -2,7 +2,11 @@
 <html xmlns:th="http://www.thymeleaf.org"
 	xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" lang="ja"
 >
-<footer th:fragment = "common_footer"> 
-<h1>Footer</h1>
-</footer>
+<div th:fragment="common_footer">
+	<footer class="footer">
+		<div class="container">
+			<p class="text-muted">Place sticky footer content here.</p>
+		</div>
+	</footer>
+</div>
 </html>
